+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add optional ability to load fixed address flashing algorithms (non PIC). (#822)
 - Added STM32WL55JCIx target. (#835)
 - Add esp32.yaml with esp32c3 variant. (#846)
+- Added target definition validation to make handling inside probe-rs easier by making some basic assumptions about the validity of the used `ChipFamily` without always checking again. (#848)
 
 ### Removed